moved user profile to a widget
